refactor: 패키지구조 변경
