--- src/main/java/nbc/devmountain/domain/chat/chatmessage/dto/request/ChatMessageRequest.java
@@ -1,5 +1,0 @@
-package nbc.devmountain.domain.chat.chatmessage.dto.request;
-
-public record ChatMessageRequest(String message) {
-
-}
