+++ src/main/java/nbc/devmountain/domain/chat/dto/ChatMessageRequest.java
@@ -1,0 +1,5 @@
+package nbc.devmountain.domain.chat.dto;
+
+public record ChatMessageRequest(String message) {
+
+}
